Connect search and reset buttons, reorganize classes a bit
