+++ naturtag/controllers/__init__.py
@@ -1,0 +1,2 @@
+from naturtag.controllers.image_controller import ImageController
+from naturtag.controllers.taxon_controller import TaxonController
